Add ansible automated node labelling for swarm managers
The platform scripts require nodes to be labelled in order to specify
stateful service locations. Eg. mongo-1 should be present on 'node-1'
the labelling needs to be done before platform deploy therefore we are
doing it in the ansible scripts.
Unfortunately, worker nodes can't label themselves. With our ansible
setup we don't have the context of the worker nodes and their labels in
one place so for now labelling them will need to be manual.

PLAT-78