Remove unnecessary files from docker image
The platform docker image is enormous and needs to be reduced if
possible.
The excluded files here save 40Mb of space - not much compared to the
1.75Gb from before.
The primary culprit for the space usage is the instant-base image.
Perhaps we neeed to look at a smaller base as well as checking there
isn't anything excessive in the instant DockerFile

PLAT-163